redid everything
