+++ ComfyNotch/Views/Settings/New/Tabs/NotchUpdatesTab/NotchUpdatesTab.swift
@@ -1,13 +1,13 @@
 //
-//  UpdatesSettingsView.swift
+//  NotchUpdatesTab.swift
 //  ComfyNotch
 //
-//  Created by Aryan Rogye on 5/24/25.
+//  Created by Aryan Rogye on 7/19/25.
 //
 
 import SwiftUI
 
-struct UpdatesSettingsView: View {
+struct NotchUpdatesTab: View {
     
     @EnvironmentObject var settings: SettingsModel
